backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "beanie>=2.0.1",
    "fastapi[all]>=0.128.0",
    "lxml-html-clean>=0.4.3",
    "motor>=3.7.1",
    "newspaper3k>=0.2.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
]
